fix: move label repository path
